get-macaroon: load defaults from config file
The script now tries to source /etc/get-macaroon.conf and ~/.get-macaroon.conf, to load site specific defaults. We wanted this to set a default IP caveat that limits access to our own systems. Defaults can always be overridden on the command line.
Additionally, the script now always shows the macaroon properties; not only when `--debug` is specified. This is to make users aware of such caveats that they haven't specified themselves.