Update check_lsc: method for getting LSC info changed
There was an error in check_lsc. It took the LSC info from a configuration.action page, like https://voms2.fnal.gov:8443/voms/dune/configuration/configuration.action. But this info can be outdated! So I updated check_lsc to first get the VOMSES info (which contains the hostname and port number that voms-proxy-init should use to verify VOMS credentials). Then, on the extracted hostname and port number, it checks with `openssl s_client` which subject and issuer DNs are returned.

This method is closer to what voms-proxy-init does and should therefore be more reliable.

I also added a --debug option, to show which curl commands are being used.