Dump HTTP header if HTTP code is not 200
get-macaroon did not provide a clear error message when something went wrong with getting a macaroon.
With this change, the HTTP header that curl received from dCache will be printed to stderr, but only if it's not 200 (success). This way, users can understand a bit more clearly what the problem was.