get-macaroon: add --fail --show-error to show all errors
The --dump-header already shows many errors, coming from the dCache WebDAV door.
But there are some errors (like "Could not resolve host") that are only shown with --fail --show-error.
So we add this. This will also show the HTTP code. But --dump-header is still useful because it shows not only the HTTP code, but also the description.